Split up filters into separate components
